requirements.txt
flask
pandas
numpy
plotly-express
matplotlib
seaborn